requirements.txt
pip~=22.3.1
wheel~=0.38.4
setuptools~=65.6.3
protobuf==3.20.0
pyairtable~=1.4.0
psycopg2-binary==2.9.3
gooddata-sdk~=1.2.0